latest compile. thermal_cam.ino now has spi buffer (not faster yet) starting to work on 128x128
starting to work on 128x128, added in spi buffer mode, but it currently is slower due to overhead and the fact that it is not yet unloading during rest of time. currently just testing it. if it is enabled everything still works.
still need to resolve artifacts issues in pixels.