32x32 support for ili9341 display
includes advanced color enhancement that increase color difference on screen. this is similar to autot emp range and to enhanced detail, however this is over a limited color range and specific to the 30x32 block the image is processing. it does not increase detail it increase color difference on screen and more specifically the color block it is in. this does mean that some intensities will be saturated, and this means it will show solid brightest color mapped.

also i modified the #21 color map so the first two colors mapped are black. this seems more appropriate than blue especially when zeroing out noise or threshold